Create BackToCluster button component
